requirements.txt
streamlit>=1.24.0
tensorflow>=2.10.0
numpy>=1.21.0
pillow>=9.0.0
matplotlib>=3.5.0
requests>=2.27.0
gdown>=4.6.0